% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{zs_server_logs}
\alias{zs_server_logs}
\title{Get the background server logs}
\usage{
zs_server_logs(n = 20)
}
\arguments{
\item{n}{Number of lines to return from the end of the log.}
}
\value{
A character vector of log lines.
}
\description{
Get the background server logs
}
\examples{
\dontrun{
zs_server_logs()
}
}
