% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{zs_stop_server}
\alias{zs_stop_server}
\title{Stop the background zeroserve server}
\usage{
zs_stop_server()
}
\value{
Logical; TRUE if server was stopped, FALSE if it wasn't running.
}
\description{
Stop the background zeroserve server
}
\examples{
\dontrun{
zs_stop_server()
}
}
