% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{zs_clear_registry}
\alias{zs_clear_registry}
\title{Clear the zeroserve resource registry}
\usage{
zs_clear_registry()
}
\value{
Logical; TRUE if registry was cleared.
}
\description{
This stops serving all currently registered resources and frees
associated memory buffers.
}
\examples{
\dontrun{
zs_clear_registry()
}
}
