% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{zs_serve_arrow}
\alias{zs_serve_arrow}
\title{Serve an object via In-Memory Arrow IPC (mori)}
\usage{
zs_serve_arrow(x, layer_id = "stream", crs = NULL)
}
\arguments{
\item{x}{A spatial object (e.g. \code{sf}, \code{duckspatial_df}).}

\item{layer_id}{A unique identifier for this data stream.}

\item{crs}{Optional Coordinate Reference System.}
}
\value{
A URL string pointing to the localhost background server.
}
\description{
Serve an object via In-Memory Arrow IPC (mori)
}
